Cast amount values to int for proper numeric comparison

+++ plugins/gangvault.php
@@ -12,7 +12,7 @@ if (array_key_exists('deposit', $_POST) || array_key_exists('deposit_points', $_
     }
     $curr = !array_key_exists('points', $_GET) ? 'money' : 'points';
     $currVault = !array_key_exists('points', $_GET) ? 'moneyvault' : 'pointsvault';
-    $_POST['damount'] = array_key_exists('damount', $_POST) && ctype_digit(str_replace(',', '', $_POST['damount'])) ? str_replace(',', '', $_POST['damount']) : null;
+    $_POST['damount'] = array_key_exists('damount', $_POST) && ctype_digit(str_replace(',', '', $_POST['damount'])) ? (int)str_replace(',', '', $_POST['damount']) : null;
     if ($_POST['damount'] === null) {
         $errors[] = 'You didn\'t enter a valid amount';
     } elseif ($_POST['damount'] <= 0) {
@@ -38,7 +38,7 @@ if ((array_key_exists('withdraw', $_POST) || array_key_exists('withdraw_points',
     }
     $curr = !empty($_POST['withdraw']) ? 'money' : 'points';
     $currVault = !empty($_POST['withdraw']) ? 'moneyvault' : 'pointsvault';
-    $_POST['wamount'] = array_key_exists('wamount', $_POST) && ctype_digit(str_replace(',', '', $_POST['wamount'])) ? str_replace(',', '', $_POST['wamount']) : null;
+    $_POST['wamount'] = array_key_exists('wamount', $_POST) && ctype_digit(str_replace(',', '', $_POST['wamount'])) ? (int)str_replace(',', '', $_POST['wamount']) : null;
     if ($_POST['wamount'] === null) {
         $errors[] = 'You didn\'t enter a valid amount';
     } elseif ($_POST['wamount'] <= 0) {
